requirements.txt
discord.py>=1.7.3,<1.8.0
python-dotenv>=0.19.0,<0.20.0
requests>=2.26.0,<2.27.0
